05-audit | make a simplier db connector need to fix a EventDate field
